Move alloc info implementation into separate module
The alloc info list's APIs need to be available to both the global
allocator and the GC. It is also not possible to define a global
anywhere *except* the top level module. It makes sense to split these in
order to have as little top level code in lib.rs as possible
